Base Classes  + Attribute Integration
+ Added a few base class stubs
   - GameBase: Should the base class for most subclasses. Provides subclasses with attribute management and notification support
   - Actor: stub, most of the player class should be moved in here

+ Add AttributeManager as parent class of GameBase
+ Changed split mod to take a list of child mods that will be given to child projectiles instead of simply passing all mods to all child projectiles.
    - Additionally, helps prevent exponential growth in number of active projectiles. Though, I have no doubt we'll add Mods that get attached to child projectiles and reintroduce the problem
+ Added initial support for equipping items.
+ Integrated attributes throughout the game. All classes inheriting from GameBase will have access to Attributes.

+++ Assets/Scripts/Enemy.cs
@@ -9,7 +9,7 @@ public class Enemy : MonoBehaviour
     private float dupeCooldown;
     public float dupeCooldownMax;
 
-    public int health = 50;
+    public float health = 50;
     public float speed = 2f;
     public GameObject player;
     private Rigidbody rb;
@@ -84,7 +84,7 @@ public class Enemy : MonoBehaviour
         }
     }
 
-    public void TakeDamage (int amount)
+    public void TakeDamage (float amount)
     {
         health -= amount;
         if (health <= 0f)
